解决启动全量校验工具问题，反向迁移分包问题，启动安装，卸载，全量迁移，全量校验进程残留问题

+++ src/main/java/org/opengauss/portalcontroller/RuntimeExecTools.java
@@ -237,9 +237,13 @@ public class RuntimeExecTools {
      * @param path Filepath.
      */
     public static void removeFile(String path) {
-        String command = "rm -rf " + path;
-        executeOrder(command, 60000);
-        LOGGER.info("Remove file " + path + " finished.");
+        if(new File(path).exists()){
+            String command = "rm -rf " + path;
+            executeOrder(command, 60000);
+            LOGGER.info("Remove file " + path + " finished.");
+        }else{
+            LOGGER.info("No file " + path + " to remove.");
+        }
     }
 
     /**
